requirements.txt
# requirements.txt
pandas
numpy
scipy
scikit-learn
requests
lightfm
python-dotenv
joblib
